Task 4: nextInt() to nextDouble()

+++ src/com/ibs/akalimullin/task4/Average.java
@@ -6,9 +6,9 @@ import java.util.Scanner;
 public class Average {
     public static void main(String[] args) {
         Scanner in = new Scanner(System.in);
-        double x = in.nextInt();
-        double y = in.nextInt();
-        double z = in.nextInt();
+        double x = in.nextDouble();
+        double y = in.nextDouble();
+        double z = in.nextDouble();
 
         double avg = (x + y + z) / 3;
         System.out.println(avg);
